move results i18n code to a package, mark surveyform i18n logic as legacy
