+++ shared/react-i18n-legacy/stringsRegistry.ts
