+++ shared/react-i18n-legacy/test/index.js
