+++ shared/react-i18n-legacy/tsconfig.json
