+++ shared/react-i18n-legacy/typings.ts
