+++ surveyform/package.json
@@ -56,7 +56,7 @@
     "@devographics/helpers": "workspace:^1.0.0",
     "@devographics/i18n": "workspace:^1.0.0",
     "@devographics/mongo": "workspace:^1.0.0",
-    "@devographics/react-i18n": "workspace:1.0.0",
+    "@devographics/react-i18n-legacy": "workspace:1.0.0",
     "@devographics/redis": "workspace:^1.0.0",
     "@devographics/templates": "workspace:^1.0.0",
     "@devographics/types": "workspace:^1.0.0",
