+++ surveyform/src/components/error/TopLevelErrorContainer.tsx.todo
@@ -7,7 +7,7 @@ import React from "react";
 
 // import NotFound from "./NotFound";
 import DefaultError from "./DefaultError";
-import { useIntlContext } from "@devographics/react-i18n";
+import { useIntlContext } from "@devographics/react-i18n-legacy";
 
 /*
 const useStyles = makeStyles({
