+++ surveyform/src/components/form/OptionLabel.tsx
@@ -7,7 +7,7 @@ import {
   QuestionMetadata,
 } from "@devographics/types";
 import { getOptioni18nIds } from "~/i18n/survey";
-import { useIntlContext } from "@devographics/react-i18n";
+import { useIntlContext } from "@devographics/react-i18n-legacy";
 
 const OptionLabel = ({
   option,
