+++ surveyform/src/components/share/ShareEmail.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useIntlContext } from "@devographics/react-i18n";
+import { useIntlContext } from "@devographics/react-i18n-legacy";
 
 const ShareEmail = ({
   subject,
